add StringTrimmer

+++ src/main/java/com/blueberry/controller/CommentController.java
@@ -4,11 +4,13 @@ import com.blueberry.model.app.AppUser;
 import com.blueberry.model.app.Comment;
 import com.blueberry.model.app.Status;
 import com.blueberry.model.dto.CommentDTO;
+import com.blueberry.model.dto.MessageResponse;
 import com.blueberry.service.AppUserService;
 import com.blueberry.service.CommentService;
 import com.blueberry.service.StatusService;
 import com.blueberry.service.UserService;
 import com.blueberry.util.ModelMapperUtil;
+import com.blueberry.util.StringTrimmer;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,6 +19,7 @@ import org.springframework.web.bind.annotation.*;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/auth/api/status")
@@ -38,7 +41,7 @@ public class CommentController {
             return ResponseEntity.notFound().build();
         }
 
-        List<Comment> commentList = status.getCommentList();
+        List<Comment> commentList = (List<Comment>) commentService.findAllByStatusIdAndIsDeleted(statusId,false);
 
         return new ResponseEntity<>(modelMapperUtil.mapList(commentList,CommentDTO.class), HttpStatus.OK);
     }
@@ -52,6 +55,7 @@ public class CommentController {
         }
         newComment.setAuthor(currentAppUser);
         newComment.setStatus(status);
+        newComment.setBody(StringTrimmer.trim(newComment.getBody()));
         newComment.setCreatedAt(LocalDateTime.now());
         Comment savedComment = commentService.save(newComment);
 
@@ -63,36 +67,41 @@ public class CommentController {
     }
 
     @PutMapping("/comments/{commentId}")
-    public ResponseEntity<CommentDTO> editCommentById(@PathVariable Long commentId, @RequestBody Comment updatedComment) {
+    public ResponseEntity<?> editCommentById(@PathVariable Long commentId, @RequestBody Comment updatedComment) {
         Comment currentComment = commentService.findById(commentId).orElse(null);
-
         if (currentComment == null) {
             return ResponseEntity.notFound().build();
         }
+        AppUser currentAppUser = appUserService.getCurrentAppUser();
+        if(Objects.equals(currentAppUser.getId(), currentComment.getAuthor().getId())){
+            currentComment.setBody(StringTrimmer.trim(updatedComment.getBody()));
+            currentComment.setUpdatedAt(LocalDateTime.now());
+            currentComment.setUpdated(true);
 
-        currentComment.setBody(updatedComment.getBody());
-        currentComment.setUpdatedAt(LocalDateTime.now());
-        currentComment.setUpdated(true);
+            Comment savedComment = commentService.save(currentComment);
 
-        Comment savedComment = commentService.save(currentComment);
-
-        return new ResponseEntity<>(modelMapperUtil.map(savedComment,CommentDTO.class), HttpStatus.OK);
+            return new ResponseEntity<>(modelMapperUtil.map(savedComment,CommentDTO.class), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(new MessageResponse("Access denied!"),HttpStatus.FORBIDDEN);
     }
 
     @DeleteMapping("/comments/{commentId}")
-    public ResponseEntity<String> deleteCommentById(@PathVariable Long commentId) {
+    public ResponseEntity<MessageResponse> deleteCommentById(@PathVariable Long commentId) {
         Comment currentComment = commentService.findById(commentId).orElse(null);
 
         if (currentComment == null) {
             return ResponseEntity.notFound().build();
         }
+        AppUser currentAppUser = appUserService.getCurrentAppUser();
+        if (Objects.equals(currentAppUser.getId(), currentComment.getAuthor().getId())) {
+            Status status = currentComment.getStatus();
+            status.getCommentList().remove(currentComment);
+            statusService.save(status);
 
-        Status status = currentComment.getStatus();
-        status.getCommentList().remove(currentComment);
-        statusService.save(status);
+            commentService.delete(commentId);
 
-        commentService.delete(commentId);
-
-        return new ResponseEntity<>("Deleted successful !!", HttpStatus.OK);
+            return new ResponseEntity<>(new MessageResponse("Deleted successful !!"), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(new MessageResponse("Access denied !!"), HttpStatus.FORBIDDEN);
     }
 }
